+++ package.js
@@ -9,7 +9,8 @@ Package.describe({
 Package.onUse( function( api ){
     configure( api );
     api.export([
-        'Bootstrap'
+        'Bootstrap',
+        'Popper'
     ]);
     api.mainModule( 'src/client/js/index.js', 'client' );
     api.mainModule( 'src/server/js/index.js', 'server' );