+++ src/client/js/index.js
@@ -4,7 +4,9 @@
 
 import '../../common/js/index.js';
 
-import '@popperjs/core/dist/cjs/popper.js';
+import * as PopperNS from '@popperjs/core/dist/cjs/popper.js';
+export const Popper = PopperNS;
+
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 import * as BootstrapNS from  'bootstrap/dist/js/bootstrap.min.js';